+++ modules/quarkus-core/src/main/kotlin/com/explyt/quarkus/core/linemarker/QuarkusBeanLibraryLineMarkerProvider.kt
@@ -83,7 +83,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
         result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val sourcePsi = uClass.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.BeanDependencies,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { findBeanDeclarations(uClass) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))
@@ -99,7 +102,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
         if (uClass.javaPsi.isMetaAnnotatedBy(QuarkusCoreClasses.INTERCEPTOR.allFqns)) return
 
         val sourcePsi = uClass.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.Bean,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(null, uClass) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -122,7 +128,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
         val psiField = uField.javaPsi as? PsiField ?: return
         if (psiField.isAnnotatedBy(QuarkusCoreClasses.PRODUCES.allFqns)) {
             val sourcePsi = uField.uastAnchor?.sourcePsi ?: return
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.Bean,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(uField.type, uField) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -135,7 +144,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
         if (!isInjectExpression(psiField) && !isLombokAnnotatedClassFieldExpression(psiField)) return
 
         val sourcePsi = uField.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.BeanDependencies,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uField) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))
@@ -162,7 +174,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
         }
 
         if (method.javaPsi.isMetaAnnotatedBy(QuarkusCoreClasses.PRODUCES.allFqns)) {
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.Bean,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(method.returnType, method) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -181,7 +196,10 @@ class QuarkusBeanLibraryLineMarkerProvider : RelatedItemLineMarkerProvider() {
     ) {
         for (uParameter in method.uastParameters) {
             val sourcePsi = uParameter.uastAnchor?.sourcePsi ?: continue
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.BeanDependencies,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uParameter) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))