+++ modules/quarkus-core/src/main/kotlin/com/explyt/quarkus/core/linemarker/QuarkusBeanLineMarkerProvider.kt
@@ -91,7 +91,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
         result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val sourcePsi = uClass.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.BeanDependencies,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { findBeanDeclarations(uClass, module) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))
@@ -108,7 +111,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
         if (uClass.javaPsi.isMetaAnnotatedBy(QuarkusCoreClasses.INTERCEPTOR.allFqns)) return
 
         val sourcePsi = uClass.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.Bean,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(null, uClass, module) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -145,7 +151,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
         val psiField = uField.javaPsi as? PsiField ?: return
         if (psiField.isAnnotatedBy(QuarkusCoreClasses.PRODUCES.allFqns)) {
             val sourcePsi = uField.uastAnchor?.sourcePsi ?: return
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.Bean,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(uField.type, uField, module) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -177,7 +186,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
         result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val sourcePsi = uField.uastAnchor?.sourcePsi ?: return
-        val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+        val builder = NavigationGutterIconBuilder.create(
+            QuarkusCoreIcons.BeanDependencies,
+            QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+        )
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
             .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uField, module) })
             .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))
@@ -204,7 +216,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
         }
 
         if (method.javaPsi.isMetaAnnotatedBy(QuarkusCoreClasses.PRODUCES.allFqns)) {
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.Bean)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.Bean,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithInject(method.returnType, method, module) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.autowired.candidate"))
@@ -224,7 +239,10 @@ class QuarkusBeanLineMarkerProvider : RelatedItemLineMarkerProvider() {
     ) {
         for (uParameter in method.uastParameters) {
             val sourcePsi = uParameter.uastAnchor?.sourcePsi ?: continue
-            val builder = NavigationGutterIconBuilder.create(QuarkusCoreIcons.BeanDependencies)
+            val builder = NavigationGutterIconBuilder.create(
+                QuarkusCoreIcons.BeanDependencies,
+                QuarkusCoreBundle.message("explyt.quarkus.gutter.group.bean")
+            )
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
                 .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uParameter, module) })
                 .setTooltipText(QuarkusCoreBundle.message("explyt.quarkus.gutter.tooltip.title.choose.bean.candidate"))